Added BRAKE option to AVOID_ADSB mode.
With this option selected, an imminent potential collision will cause
the vehicle will switch to rapidly descend to a safe altitude and
then switch to BRAKE mode.

Parameters were added to toggle whether the vehicle should first descend
to a safe altitude prior to switching to brake, specify the safe altitude,
and specify the descent rate.

This commit is dependent on changes to mavlink.

A warning is send upon startup if the avoid brake altitude is lower than the
avoidance minimum altitude. If that is the case, then the vehicle
will descend until it reaches the avoidance minimum altitude and then
will stay in ADSB_AVOID mode instead of switching to BRAKE.